backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
Pillow==10.2.0
PyPDF2==3.0.1
google-generativeai==0.3.2
numpy==1.26.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0